refactor(Avatar): Make avatar more versatile
Avatar can now take in a size property, and
refactored the styles to not be absolute positioned.
This will help with it being placed in other components.
Cards Primary component also was affected by this change,
and was fixed accordingly.

rid of unneeded comments

fix

docs conflict

now build

fixed merge

fix docs