feat(Portal): Add more functionality to portal
This reworks the portal class to give it a little more utility in the
specific use case of a drawer, it might also be useful outside of the
drawer context but in most cases the 'overlay' mode would be used.

First, it uses injectGlobal to inject some global styles that we need in
order for the shift portal mode to work. Shift uses flex box positioning
to "fake" the positioning of the drawer by having flex on the body we
can use :before and :after to add flex items into the body and reserve
space for absolutely positioned items.

It also moves Overlay component into its own file and creates a new class
for the Shift component.