chore(Dialog): Update styling
The previous iteration of the dialog component had adjusted
smc-portal-container to be 100% width and height in order to use flex
appropriately. Unfortuantely this intefered with the click - out effect
on the drawer that is using the native events from the Overlay
component.

In the future i'd like to investigate adding flex to the overlay as well
as positioning props to abstract some of the dialog's styles and prevent
the need for a lot of work for each individual component that uses the
portal.