fix(inkbar): fix inkbar flash
On load the inkbar flashes to 100% width before shrinking to the
correct width. This is because it passed a value of undefined for the
tabWidth prop. This commit defaults that value to 0 for the tab, so
that it will not display the inkbar until after the width has been
determined.