chore(update patterns): Remove classNames
Goal: Remove the need to create our own components simply to attach a
class name, or other identifiable attribute.

Secondary Goal: Remove classnames, which can cause styling issues, and
instead use data-smc attribute to identify the component name. Where
necessary we should use the component itself in interpolation to
actually style sub components.

This just removes the custom button component. Wrapping a component in
'styled' removes the automatic whitelisting of props that react does.
This removes the need to declare an intermediate click action and white
list the props on the button.