fix(Card): Fix hover warning
React was throwing a warning on passing a non boolean
attribute to the component. This fixes it by not passing
it through the CardComponent, and only the Card.

Note: This also has a slight fix to the import statements of
Image in Avatar, as well as, Circle.

BREAKING CHANGE: This does break some styles on the Card component