feat(Appbar): Add Navigation Options
This adds the ability to add a Navigation Menu to the Appbar component.
It has minor state itself to keep track of the position of the inkbar,
and has a couple hooks for updating the position of the inkbar based on
the resizing of the browser or the active item changing.