feat(Toolbar): Implement Toolbar Spec
This changes the Appbar component to be a toolbar. Appbars are just
specific implementations of a toolbar. This is largely to spec but has
some work to get to full spec. Main things that are missing right now is
the ability to change the padding / margins around icons and whatnot in
the toolbar when in mobile views.

More work to come on this but for now this is a MVP approach.