refactor(Contexts): Use official React 16.3 api
This removes the polyfill in favor of the real thing. There is a
breaking change in that I have reverted the 'functional hoc' style for
the withScreenSize and withRowState contexts because it comes with side
effects that are unintentional with other apis that were exposed by
react 16.3 update. Namely this means the forwardRef api.

BREAKING CHANGE:
Usage of withScreenSize/withRowState must be updated.