fix(transition): Fix Mac import issue
If platform.os.family is falsy, fallback to ''. This is important
because, when running jest in projects that import this project,
running platform.os.family.match() was throwing a type error on Mac
systems which have this as 'null' during the tests (and this happens
before the jest setup file can mock platform.os.family). This causes
the import to error and tests to fail. This is not an issue on Linux
systems because they are passed 'Linux' for platform.os.family.